Replace single_test with runner
Adds support for "test_env" to the runner script
