+++ test/integrationtests/skills/runner.py
@@ -25,6 +25,9 @@ developer's tests.
 import glob
 import unittest
 import os
+from os.path import exists
+import sys
+import imp
 import argparse
 from test.integrationtests.skills.skill_tester import MockSkillsLoader
 from test.integrationtests.skills.skill_tester import SkillTest
@@ -40,7 +43,8 @@ parser = argparse.ArgumentParser(description=desc)
 parser.add_argument("skill_path", nargs='?', default=os.getcwd(),
                     help="path to skill to test, default=current")
 args = parser.parse_args()
-HOME_DIR = args.skill_path
+HOME_DIR = os.path.dirname(args.skill_path + '/')
+sys.argv = sys.argv[:1]
 
 
 def discover_tests():
@@ -49,9 +53,12 @@ def discover_tests():
     For all skills with test files, starten from current directory,
     find the test files in subdirectory test/intent.
 
-    :return: skills and corresponding test case files found
+    Returns:
+        Test case files found along with test environment script if available.
     """
+
     tests = {}
+    test_envs = {}
 
     skills = [HOME_DIR]
 
@@ -66,19 +73,34 @@ def discover_tests():
         if len(test_intent_files) > 0:
             tests[skill] = test_intent_files
 
-    return tests
+        # Load test environment script
+        test_env = None
+        if exists(os.path.join(skill, 'test/__init__.py')):
+            module = imp.load_source(skill + '.test_env',
+                                     os.path.join(skill, 'test/__init__.py'))
+            if hasattr(module, 'test_runner') and callable(module.test_runner):
+                test_env = module
+        test_envs[skill] = test_env
+
+    return tests, test_envs
 
 
 class IntentTestSequenceMeta(type):
     def __new__(mcs, name, bases, d):
-        def gen_test(a, b):
+        def gen_test(a, b, test_env):
             def test(self):
                 t = SkillTest(a, b, self.emitter)
                 if not t.run(self.loader):
                     assert False, "Failure: " + t.failure_msg
-            return test
 
-        tests = discover_tests()
+            def test_env_test(self):
+                assert test_env.test_runner(a, b, self.emitter, self.loader)
+
+            if test_env:
+                return test_env_test
+            else:
+                return test
+        tests, test_envs = discover_tests()
         for skill in tests.keys():
             skill_name = os.path.basename(skill)  # Path of the skill
             for example in tests[skill]:
@@ -86,7 +108,9 @@ class IntentTestSequenceMeta(type):
                 test_filename = os.path.basename(example)
                 test_name = "test_Intent[%s:%s]" % (skill_name,
                                                     test_filename)
-                d[test_name] = gen_test(skill, example)
+                test_env = test_envs[skill]
+                d[test_name] = gen_test(skill, example, test_env)
+
         return type.__new__(mcs, name, bases, d)
